requirements.txt
click==6.6
Django==1.10
django-crispy-forms==1.6.0
django-filter==0.14.0
djangorestframework==3.4.6
drfdocs==0.0.11
httpie==0.9.6
Jinja2==2.8
livereload==2.4.1
Markdown==2.6.6
MarkupSafe==0.23
mkdocs==0.15.3
mkdocs-bootstrap==0.1.1
mkdocs-bootswatch==0.4.0
psycopg2==2.6.2
Pygments==2.1.3
PyYAML==3.11
requests==2.11.1
six==1.10.0
tornado==4.4.1
